# docs/openapi.yaml
openapi: 3.1.0
info:
    title: 'Uvision API Documentation'
    version: 1.0.0
    description: 'Comprehensive API documentation for Uvision Backend'
    contact:
        name: 'API Support'
        email: support@uvision.com
    license:
        name: MIT
        url: 'https://opensource.org/licenses/MIT'
servers:
    -
        url: 'http://localhost:3000'
        description: 'Development server'
    -
        url: 'https://api.uvision.com'
        description: 'Production server'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'JWT authorization header using the Bearer scheme.'
        ApiKeyAuth:
            type: apiKey
            in: header
            name: x-access-token
            description: 'API key for accessing protected endpoints'
    schemas:
        Error:
            type: object
            properties: {status: {type: integer, example: 0}, message: {type: string, example: 'Error message'}, error: {type: object}}
...
        User:
            type: object
            properties: {id: {type: integer, example: 1}, email: {type: string, format: email, example: user@example.com}, name: {type: string, example: 'John Doe'}, profilePicture: {type: string, example: 'https://example.com/profile.jpg'}, createdAt: {type: string, format: date-time}, updatedAt: {type: string, format: date-time}}
...
        LoginRequest:
            type: object
            required: [email, password]
            properties: {email: {type: string, format: email, description: 'User''s email address', example: user@example.com}, password: {type: string, format: password, description: 'User''s password', example: password123}}
        SignUpRequest:
            type: object
            required: [email, name, password]
            properties: {email: {type: string, format: email, description: 'User''s email address', example: newuser@example.com}, name: {type: string, description: 'User''s full name', example: 'John Doe'}, profilePicture: {type: string, description: 'URL to user''s profile picture', example: 'https://example.com/profile.jpg'}, password: {type: string, format: password, description: 'User''s password', example: password123}}
        ChangePasswordRequest:
            type: object
            required: [currentPassword, newPassword, confirmPassword]
            properties: {currentPassword: {type: string, format: password, description: 'User''s current password', example: oldpassword123}, newPassword: {type: string, format: password, description: 'User''s new password', example: newpassword123}, confirmPassword: {type: string, format: password, description: 'Confirmation of new password', example: newpassword123}}
        AuthResponse:
            type: object
            properties: {status: {type: integer, example: 1}, message: {type: string, example: Success}, data: {type: object, properties: {user: {$ref: '#/components/schemas/User'}, token: {type: string, description: 'JWT access token', example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...}}}}
...
    responses:
        UnauthorizedError:
            description: 'Authentication information is missing or invalid'
            content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {status: 0, message: 'Unauthorized access'}}}
...
        ValidationError:
            description: 'Validation error in request data'
            content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {status: 0, message: 'Validation failed', error: {field: 'Field validation error message'}}}}
security:
    -
        bearerAuth: []
    -
        ApiKeyAuth: []
...
paths:
    /auth/signup:
        post:
            summary: 'Register a new user'
            description: 'Create a new user account with email, name, and password'
            tags: [Authentication]
            security: []
            requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SignUpRequest'}}}}
            responses: {'200': {description: 'User successfully registered', content: {application/json: {schema: {$ref: '#/components/schemas/AuthResponse'}}}}, '400': {$ref: '#/components/responses/ValidationError'}, '409': {description: 'User already exists', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {status: 0, message: 'User already exists with this email'}}}}}
    /auth/login:
        post:
            summary: 'User login'
            description: 'Authenticate user with email and password'
            tags: [Authentication]
            security: []
            requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginRequest'}}}}
            responses: {'200': {description: 'Login successful', content: {application/json: {schema: {$ref: '#/components/schemas/AuthResponse'}}}}, '400': {$ref: '#/components/responses/ValidationError'}, '401': {description: 'Invalid credentials', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {status: 0, message: 'Invalid email or password'}}}}}
    /auth/verifyToken:
        get:
            summary: 'Verify JWT token'
            description: 'Validate the provided JWT token and return user information'
            tags: [Authentication]
            security: []
            parameters: [{in: header, name: x-access-token, required: true, schema: {type: string}, description: 'JWT token to verify', example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...}]
            responses: {'200': {description: 'Token is valid', content: {application/json: {schema: {type: object, properties: {status: {type: integer, example: 1}, message: {type: string, example: 'Token is valid'}, data: {type: object, properties: {user: {$ref: '#/components/schemas/User'}, tokenValid: {type: boolean, example: true}}}}}}}}, '400': {description: 'Token missing or malformed', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}, '401': {$ref: '#/components/responses/UnauthorizedError'}}
    /auth/changePassword:
        post:
            summary: 'Change user password'
            description: 'Change the password for an authenticated user'
            tags: [Authentication]
            security: [{bearerAuth: []}, {ApiKeyAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChangePasswordRequest'}}}}
            responses: {'200': {description: 'Password changed successfully', content: {application/json: {schema: {type: object, properties: {status: {type: integer, example: 1}, message: {type: string, example: 'Password changed successfully'}, data: {type: object}}}}}}, '400': {description: 'Password validation error', content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {status: 0, message: 'Current password is incorrect'}}}}, '401': {$ref: '#/components/responses/UnauthorizedError'}}
...